feat: props 정리 및 변수 제거
