+++ src/advanced/ui/index.ts
@@ -4,3 +4,4 @@ export * from './cart';
 export * from './coupons';
 export * from './payments';
 export * from './header';
+export * from './icons';
